Fix out of bounds read when receiving a StechecPkt
The StechecPkt packet type is handled a bit differently from the others packet
because it has a variable length (`nb_args` integers). Thus its size is also
variable. This variable size was correctly handled on the network receive/send
side, but when doing a copy of the packet object, the default copy constructor
was used and the number of arguments was not respected during the copy.

This patch fixes the problem by adding a specialized copy constructor to the
StechecPkt class, which only copies `nb_args` arguments from the packet.