Allow sending part of a packet
There seems to be edge cases (when the socket TX buffer queue is full, for
example) where all data can not be sent in a single call to send(2). Instead of
disconnecting the client in this case, simply resend what have not been sent.

I'm not sure of the exact implications of this commit on the receiving side,
the data should be checked for consistency.