Remove the obsolete DirectCx from the code base
The DirectCx is a direct client-to-client communication channel which does not
use multiple processes, but instead only one process with multiple threads and
a shared message queue.

This is not used at all at Prologin and most of the code is buggy or commented
out. This commit removes this obsolete code from the project.