Remove the vs.net project files
Stechec does not build with VS.NET in its current state, and the provided
project files are unmaintained, buggy and old. Remove them.

Note that there are almost valid Code::Blocks project files used in the
Bombazero fork of Stechec. If it is needed, we could import them with a few
modifications.