Search for Python in the configure script
Since the Prologin 2011 GUI will be written in Python, we need to integrate
Python search to the current autotools-based build system.

This commits adds a new file, python.m4, which defines a macro used to search
for a Python 2.x binary (python or python2 depending on the system). This macro
is then used in configure.ac when required.