Added the merge operation in InternalTraineeMoto
This operation is not tested yet: tests would need corresponding Api and Action
implementation.

It leaves the second moto empty (with no more position), so it needs to remove
the second moto from the GameData moto’s list… without changing all the moto’s
ID’s. That’s why we have to convert this list (currently a std::vector) to a
std::map, in order to keep the moto’s ID’s even with deletions. That will be
done in a future commit… Stay tuned!