Added Api/Action for couper and fusionner. Cleaned unpacketting.
Added empty implementations for couper_trainee_moto and fusionner in Action and
Api (they need to be linked to the corresponding functions in
InternalTraineeMoto).

Instead of handling each Packet type in Action::act_from_pkt, a static methods
"recevoir" have been added to every Action* classes in order to build an action
from a packet. It reduces the size of Action::act_from_pkt and split its
independant processings.