Rename meta-server to worker-node
Meta server is kind of a stupid name for the servers handling the compilation
and match running jobs. There is nothing meta to this, it's only simple work.
Thus, this commit renames meta-server to worker-node, in order to have a
consistent master/worker analogy (master/meta does not mean anything).