Added basic heartbeat functionality to the worker node
The worker node is now able to send a recurrent heartbeat to the configured
master node in order to make it aware of its existence. Currently, the
heartbeat contains no meaningful data, but it should soon include things like
the worker hostname/port couple as well as the number of available job slots.