Add a default configuration file for the worker-node
Some configurable values must be defined for the worker-node, including where
to connect (master host/port), a shared secret to use when communicating with
the master in order to sign requests with hmac, and the number of available
jobs slots on the worker.

This will also include the path to the shell scripts used for the several
worker jobs, like the match launcher or the AI compiler.