Generate the worker Makefiles directly in the specified directory
When generating worker Makefiles using "generator server", the Makefiles are
created in $install_path/makefile instead of $install_path. In most cases it is
useless as the user will specify a non existing directory as the installation
path.

This commit modifies the behavior of the generator and generates Makefiles
directly in the installation path.