Remove the forbidden functions files
The forbidden functions mechanism is a bad security system which can be avoided
so easily that it had better not be there in the first place. At least, it will
not give the impression that Stechec is secure.

Moreover, these forbidden functions include some string manipulation functions
which should be completely fine (but people were not able to use them during
last year finals), and sometimes pass a too long command line to the dynamic
linker.