Rename stechec_lime to stechec_main
The entry point of a Stechec champion is usually in includes/main.cc, but when
compiling a champion on a worker node this file must have a different name in
case the contestant is already using a file named "main.cc". However,
stechec_lime is... lame (sorry for the bad joke), and I think stechec_main
would be a better name for it.