Fixed the GameData::current_player bug
The field GameData::current_player used to be initiallized to 0, and to be
turned to 1 at the first turn, whereas the first team to play is the team
0. This field is now initiallized to (MAX_PLAYER - 1) (but there’s no such
constant yet, so 1 is used instead).

Amazingly, the action fields checking was using !GameData::get_current_player()
in order to check that the current player wasn’t moving the other team’s motos,
and that explains why this bug did not bother the game previously. That’s also
fixed now.