Added 'my_config.ini' to the .gitignore
The config.ini file is a kind of pattern for all of us, but instead of modifying
it, I created my_config.ini in order to locally customize it. I’m fed up with
seeing it when I run 'git status', therefore I want to add it to the .gitignore.

Any question?