Added methods to GameData that computes score increases and potential evolving.
The potential of a source decreases Nth times when N trainees_moto bind it to at
least one complementary source by trainee_moto. Each potential’s polarity is
handled symmetrically.

The score of the trainee_moto’s owner is increased by the sum of the sources’
absolute potentials that it binds only if both polarity are bound to the
trainee_moto.