Changed the Actions API to separate serialiser/envoyer
The envoyer method is now generic and is used for all actions. To get the
informations it needs to send, it calls the type() and serialiser() methods
which returns respectively the action type and the parameters.

More importantly, the API is now able to serialize actions to give them to the
GUI.